Hi list,
I've implemented embedder support, i.e. the host part, for st. This
allows clients to embed into the st window and is useful if you start X
applications from the terminal. For example:

$ surf -e $WINDOWID

The behavior is similar to Plan 9 where applications can take over
windows.. as far as this is possible in X ;).

The attached patch is against git master and I intent to put it into the
wiki unless someone things it's worth for mainline.

Cheers Jochen

Source: https://lists.suckless.org/hackers/2001/17072.html